+++ package.json
@@ -60,7 +60,6 @@
     "graphql": "^15.5.0",
     "graphql-tag": "^2.12.3",
     "iso8601-duration": "^1.3.0",
-    "lodash": "^4.17.21",
     "lodash-es": "^4.17.21",
     "rxjs": "~6.6.7",
     "tslib": "^2.2.0",
@@ -102,6 +101,7 @@
     "jest-html-reporter": "^3.2.0",
     "jest-junit": "^12.0.0",
     "jest-preset-angular": "^8.4.0",
+    "lodash": "^4.17.21",
     "ng-mocks": "^11.10.0",
     "ng-packagr": "^11.2.4",
     "prettier": "^2.2.1",
